pyproject.toml
[project]
name = "sql-debug-env"
version = "1.0.0"
description = "OpenEnv environment for training AI agents to debug SQL queries"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "openenv-core>=0.2.0",
    "websockets>=12.0",
    "PyYAML>=6.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.openenv]
name = "sql-debug-env"
version = "1.0.0"
